Not exactly pleased with this.
To ensure fonts have a solid block char, right now it needs to draw a 3x3 white square on the last Texture used by a Font. The problem is, Texture can't be written to unless it is "unmanaged" and to make it unmanaged, we need to create it from a Pixmap. This breaks duplicates atlas pages and is generally not great...