Get TypingLabel size set before typing starts.
I thought this would be complicated, but, it actually wasn't bad at all. Now the markup() call applies to both the layout and the workingLayout (the first with a wrap width, the other without), which measures the line heights and such all at the very start. Then, we only render up to the current "cursor position" that glyphCharIndex specifies. We no longer need to edit workingLayout so often, and it may be possible to limit its changes to those from effects.


Former-commit-id: 38e2742f5b0d307de0856c79e58da840d8a785fd