Allow all of Twemoji to be used.
Before, having both the emoji chars and the name as words filled up all of the Unicode private use area before the atlas was complete. By aliasing the emoji and name to the same GlyphRegion, this halves the space added to mapping, and allows all of the Twemoji atlas to be used.


Former-commit-id: cd04104bb80167334bf6288c922dcaff47d54d60