Switch to getters/setters for TextraLabel.font.
The setter defaults now to calling regenerateLayout(), which would otherwise be an unintuitive extra step that would need to be done if the font changed. Some code may change as a result of font becoming protected; you can still access it using getFont() and set it without regenerating using setFont(font, false).


Former-commit-id: 04a0c3f75ba8f4613f62ddb838f45b8b8be458ab