Make earlier wrap behavior automatic, and...
When wrap is set to false, this should do a few steps (at least for TypingLabel) to allow text to continue up to the layout's target width. Manual newlines are always respected unless the max lines have already been reached.

The main bug I found in the tests was that table.pack() wasn't being called after changing the table (or just one Cell in it, really). Without this, all sorts of strange behavior started happening.


Former-commit-id: 35fb79e9ede5afe35d2b1fd1f6a2396c782b3624