Change bleep mode to jostle; remove old SmallCaps.
That is, the TypingLabel-only Effect for small caps is gone, because it didn't work as well as the mode in Font. The change from bleep to jostle just means that it works on more than letters, and while it won't change the graphic, it still moves around the letters once (not continuously).


Former-commit-id: d031a0c336528037e0c7b46ee7ec53ff1a422845