Bizarre, mostly-complete fix for emoji positions.
Because scaleX and scaleY were calculated differently for emoji, variables that depended on parts of their old values, like cellHeight, acted strangely. That one depends on font.scaleY, and that is normally incorporated into scaleY, but isn't for emoji. I still need to fix underline and strikethrough's usage of originalCellHeight (and Width, maybe?), because those can't be translated to be equivalent to cellHeight if scaleY means something different.