Sudden celebratory commit! Needs cleanup.
I think this gets the exceptionally-tricky underline/strikethrough code over emoji and other images to work, at least for the case where it was totally broken before (when the underscore or dash char is used from the font). It still probably needs some work for the more common box-drawing case.


Former-commit-id: 45dcf9ac6da7b8384944fa94e74da3f0c7c205bd