Little adjustments to reduce wobbliness.
Fonts seem to have inconsistent baselines when they have fractional positions that aren't small multiples of powers of two. That is, 7.0f/8.0f scale works fine, but 4.0f/5.0f doesn't. This has to do with how floats are on computers, I guess; 0.8f isn't exactly-representable.

This also changes the height and line height on OpenSans if you load it from KnownFonts.


Former-commit-id: bbc348435095187789c70697f52d68f162ad89ff