We may not need a custom Font for ColorfulBatch!
Each of the ColorfulBatch classes has a special case for normal Sprites, which have 20 floats per array of sprite vertices. This seems to be enough to render a Font without needing a special subclass. However, the positions here are all wrong, and the black outline is actually purple, so I may have some work to do.


Former-commit-id: 2f2d3cc5dba4905ce0ee74f9f89599c58054bba9