Per-glyph drift is gone for box-drawing. But...
It still isn't aligned correctly to other characters, such as the characters an underline is supposed to be under. The underlines are correct relative to each other, at least. This happened because I started offsetting the origin, effectively, so it actually is at the center of what we're rotating. I thought this already was happening in regular glyphs, but apparently not.


Former-commit-id: 1df3430ff2309ce5f121ee1b705faa4a78557048