Change how lowering bold strength looks.
This wasn't ever done before I added LanaPixel, so I hadn't noticed how much it helps pixel fonts. It did need some changes to how it assigns the left vs. the right bold expansion, but only for boldStrength between 0 and 1 exclusive. LanaPixel is definitely broken with integer positions on, but this seems to be a bug with integer positions.


Former-commit-id: dcc5c84431fc4dc4b53441c0551c770bf59053a9