Woohoo! The message box displays well.
Big thanks to both Winter and raeleus for suggesting ways to debug this insanity. It turns out the messageGroup had width = 0 because it was sized with root.getWidth(), and all of the TypingLabel items are width 0 because they wrap. Setting a width that won't be 0 seems to be most of what it needs. Once this is done for messageGroup, it also needs to be done for each Container.


Former-commit-id: 017c6fb83ef4d8a670396470e8305da7eab94802