Get larger lines to display correctly.
This also avoids what seems to have been a bug with the heights of Lines in a Layout. If a Line was taller than normal, it would also make the next line at least the height of the last glyph in the taller Line. This made the other bug harder to see, but it was still there.


Former-commit-id: 11ddf781c06e40adbe3c2f9c6fbadd7757f14cbe