Bring back loadUnicodeAtlas().
It's needed, at least for now, because if a user of a game (that uses TextraTypist) specifies a file encoding other than UTF-8, loading the emoji atlas would not work at all. If file encoding were unspecified, it might work, and if specified as UTF-8, it would be likely to work. It doesn't seem to matter what the game specified as its file encoding in the previous commit; that would have to be UTF-8 to use emoji in Strings anyway. The user's platform encoding was still taking precedence in some way. Now it is overruled and uses UTF-8, at least for the emoji atlas.


Former-commit-id: e7facdebb0d6347f145a6761792f75c8cff0c61e