Add Issue16Test by lucas-kakele, use his fix.
That is, invalidateHierarchy() probably should be called in TextraLabel.layout(), and I don't know if TypingLabel needs the same thing, something similar, or nothing like this at all.

This also exposed a really strange bug when a scaled BitmapFont is passed to the Font constructor; the positions get thrown way off. Constructing the Font with an unscaled BitmapFont, then scaling the Font, works fine.


Former-commit-id: 6cfa56e7dca49c6503841b3a0e67352d57b7e394