It mostly works!
The vertical position and size of glyphs were both insane to get right, but they're fine now (oh, and y-offset, too). Kerning seems to be fine even with SDF fonts. When a font is scaled up, it seems to jump up from the baseline quite a lot, which might be fix-able.


Former-commit-id: 3d166aacc2eaaaa90b29fa1ddf31e4fd69f4acda