My fork of msdf-atlas-gen does the trick!
It's all chlumsky's work in my fork, because all I did was set the new spacing variable that can't be changed in the CLI yet. But, with spacing of 1 instead of 0 or -1, all the filtering artifacts go away. My fork is at https://github.com/tommyettinger/msdf-atlas-gen and it hopefully won't be needed once the main version allows spacing to be set.


Former-commit-id: eb4fc39f6c5932b29d4ba1f4ecc70e5fe9e2309e