Test and refine ZipperEffect.
The speed parameter for most of the existing effects seems to be inverted, making the effect slower when speed is higher. I'm debating whether to make the newer effects match that behavior and to rename speed to "extend" or something. I might also change old effects, though that's a breaking change... Less likely.