Commence th' breakin'!
The goal here is to be able to scale all emoji per-font to match the characteristics of that font, not per-glyph, which wasn't easily configurable before anyway. It also made the x advance unusable as an actual horizontal advance. This doesn't resolve most of this yet... Still working. Any configuration of xAdvance for inline images (like atlases and emoji) will need to change, though it's now runtime-adjustable easily. It could in theory be user-adjustable, or even adjusted by an effect...