Complex line-height fix.
This has a few parts. It makes the added newline and space chars produced by wrapping actually use the current settings for line-height, etc. That makes line-height calculations not change the line height to at least 100% because they encounter a 100% height space or newline. Second, line.height includes scale when calculateSize() processes it. Third, the movement of the first line is no longer different from the rest in TextraLabel; I may need to investigate this in TypingLabel next. Lastly, osy, a really weird compromise variable, now no longer factors in scale for some reason. It fixes the descenders, I'm not complaining.