All right, this works.
With different stretch factors, it maintains the center position of the emoji. Existing un-stretched emoji/images are unchanged. The controller buttons do need offsetX and offsetY to be -0.5 * the width or height of an image in the atlas (each of which is about 100x100). I don't know what other atlases will need.