Partial improvement to underline?
There are issues when the ratio of x-height to cap-height is different from expected. In Grenze and Yataghan, x-height and cap-height are very close, and the underline is too low. In Tangerine, the opposite is true, and strikethrough is too high. There's nothing stopping me from manually adjusting the underline metric in the original JSON files, except I would need to be able to compress them again eventually. I could even add a strikethrough metric, if needed.