Fix some issues in EffectSpeedCheck.
The Skip button would immediately restart because the end() event was set to restart upon completion. Rebuilding would mess up the MSDF shader somehow; this was fixed by calling resizeDistanceField() with the current viewport size and the current viewport when we recreate the label. There also weren't enough lines in the window or the Layout to show all effects!