Try to accommodate BitmapFont's lousy padding.
As I have said for years, BitmapFont should not be using the padding values for any rendering purpose, but here we are. Now `new BitmapFont()` and `new Font()` should be identical except for wrapping, where the BitmapFont keeps the padding information and changes how it wraps depending on it, and Font does not.