Complex fix! #27 is complete!
(I think, anyway...) The solution here was to make newlines visible, but zero-width (since we already had code to add zero-width spaces, this was easy). This allowed the code that wraps when a newline is read in to be moved to after the check for too much width being shown, so the example will wrap correctly when the long span of `q` exceeds the boundary, and only then will `\n` be read in on the next char.