Justify settings are starting to work. 11 years!
I have now committed at least once a day, every day, without stopping, for 11 years. But as for justify settings, they need to get fixed in a number of places. While the text is typing, they jitter back and forth between some wider size and a normal one (with mostly-correct justification). Any size changes in the markup aren't used when calculating the line width, which is wrong. Underline position also seems quite wrong for at least the larger text in the test...