More work is needed in regenerateLayout().
Until now, it still pulled scale as a 4-bit int from each glyph, which yields very wrong sizes in the current code. But, we can't pull from layout.advances yet because regenerateLayout() changes the size of glyphs independently of advances (which is a bug). I need to somehow wrap changes to a Layout or any Line in it so that advances, offsets, scales, and rotation all update at the same time when a glyph is added or removed...