Update TypingLabelTest with new syntax.
We had used the old black outline syntax in one place, so it continued after it should have ended. There were several other cases like this. The origins for rotation seem slightly off to the right for variable-width text, which is probably related to a misuse of cellWidth when it shouldn't be used for non-cellular fonts.