Change to KnownFonts dat and json.lzma files.
The .dat takes longer to read than the .json.lzma, and both take longer than the .fnt, but still both are under half a second to load 300 Fonts for 300 labels. The .json.lzma and .dat both look better when unscaled, but the line height seems wrong when scaledHeight is set.