This edit to OpenSans30.fnt is quick to do.
I might be able to provide a constructor argument (another one...) or static variable that affects how padding is modified when loaded from a BitmapFont. I don't like that padding is considered at all by libGDX when it creates BitmapFonts, since it runs counter to the file format spec, but because BitmapFont uses it, when we load a BitmapFont it has to be at least considered a little.