Reproduce #36 ; thanks. Lucas, for the easy test.
I put the assets into a different directory because we already have a folder called `crash/` and I didn't want crash.atlas and crash.png to get mixed up with it. I also changed the background to gray and set the texture filter to linear, which doesn't help the crash unless the ellipsis is disabled. Without the ellipsis, this works well.