We should be able to load .fnt files from subdirs!
I thought I had fixed this a long time ago... If the image isn't specified, the .fnt will look up a path to an image that should be pretty much always next to the .fnt; otherwise it should be specifically given as a TextureRegion to the constructor (which is the most flexible solution). This doesn't handle relative paths inside .fnt files, but because the .fnt is essentially a unit with its .png file and means nothing without it, they are often copied into subfolders together, and this should still work!